Reorganize repository structure to support multiple templates
