--- res/.git.bat
@@ -1,13 +1,0 @@
-@echo off
-:: The code is using gitversion so it cannot be build without git commits.
-:: This script will initialize this folder as git repository and add "Initial commit". 
-
-:: initialize git repository with main branch
-if not exist .git git init --initial-branch=main
-:: check if there are any commits
-git log 2>nul 1>nul
-if %errorlevel% equ 0 goto :eof
-:: add all files and commit
-git add -- .
-git commit -m "Initial commit"
-
