--- res/README.md
@@ -1,3 +1,0 @@
-# Janda.Go
-
-Console application startup code was created by Janda Go.
