+++ src/Console.Go/.github/workflows/build.yml
