+++ src/Console.Go/.gitignore
@@ -348,3 +348,7 @@ MigrationBackup/
 
 # Ionide (cross platform F# VS Code tools) working folder
 .ionide/
+
+# Exclude Janda.Go windows batch scripts
+.git.bat
+.publish.bat
