+++ src/Console.Go/README.md
@@ -1,0 +1,10 @@
+# ConGo
+
+Created from [Janda.Go](https://github.com/Jandini/Janda.Go)
+
+
+
+
+
+
+
