+++ src/Console.Go/src/ConGo.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.0.32014.148
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Janda.Go", "Janda.Go\Janda.Go.csproj", "{FD237AF5-6A22-480D-8B74-C23E0A62AE61}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ConGo", "ConGo\ConGo.csproj", "{FD237AF5-6A22-480D-8B74-C23E0A62AE61}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
