+++ src/Console.Go/src/ConGo/ConGo.csproj
