+++ src/Console.Go/src/ConGo/IMain.cs
@@ -1,4 +1,4 @@
-﻿namespace Janda.Go
+﻿namespace ConGo
 {
     internal interface IMain
     {
