+++ src/Console.Go/src/ConGo/appsettings.json
